1. Basic restructuring of systests base code for added EDIFACT tests 2. Fixed code generator from schema to generate X12 ISA/IEA, added these definitions to code 2. Substituted group->class, loop->group in ESLs

+++ schemas/src/main/scala/com/anypoint/df/edi/schema/tools/SchemaToCode.scala
@@ -135,7 +135,7 @@ class SchemaDump(writer: PrintWriter) {
           case _ => acc
         }))
     val elemreps = elemrefs.filter(pair => pair._2 > 1).map(pair => pair._1).toSet
-    elemreps.toList.sortBy(element => element.ident.toInt).foreach(element => defineElement(element))
+    elemreps.toList.sortBy(element => element.ident).foreach(element => defineElement(element))
     val composites = segments.foldLeft(Set[Composite]())((acc, segment) =>
       segment.components.foldLeft(acc)((acc, comp) =>
         comp match {
